// sygec-frontend/starter/src/app/models/parametre-config.ts
export interface ParametreConfig {
    dureeMinConge: number;       // Durée minimale de congé en jours
    dureeMaxConge: number;       // Durée maximale de congé en jours
    delaiMinDemande: number;     // Délai minimal avant la date de début (en jours)
    soldeMinRequis: number;      // 
}